refactor: mv chat completion processor to biz

+++ ent/job/where.go
@@ -8,47 +8,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.Job {
+func ID(id int64) predicate.Job {
 	return predicate.Job(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.Job {
+func IDEQ(id int64) predicate.Job {
 	return predicate.Job(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.Job {
+func IDNEQ(id int64) predicate.Job {
 	return predicate.Job(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.Job {
+func IDIn(ids ...int64) predicate.Job {
 	return predicate.Job(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.Job {
+func IDNotIn(ids ...int64) predicate.Job {
 	return predicate.Job(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.Job {
+func IDGT(id int64) predicate.Job {
 	return predicate.Job(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.Job {
+func IDGTE(id int64) predicate.Job {
 	return predicate.Job(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.Job {
+func IDLT(id int64) predicate.Job {
 	return predicate.Job(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.Job {
+func IDLTE(id int64) predicate.Job {
 	return predicate.Job(sql.FieldLTE(FieldID, id))
 }
