+++ ent/schema/request_execution.go
@@ -29,8 +29,8 @@ func (RequestExecution) Indexes() []ent.Index {
 
 func (RequestExecution) Fields() []ent.Field {
 	return []ent.Field{
-		field.Int("user_id").Immutable(),
-		field.Int("request_id").Immutable(),
+		field.Int64("user_id").Immutable(),
+		field.Int64("request_id").Immutable(),
 	}
 }
